Move the e2e histograms code to somewhere where it can be exported
